projects/Clinisight/requirements.txt
openai
python-dotenv
beautifulsoup4
lxml
requests
fastapi
uvicorn